- made the client work - failed due to date vars bing passed around and over writted  - better naming convention per module that handles any changes to default data

+++ iptables/killswitch.py
@@ -10,6 +10,7 @@ This file kills the app when user enters CTRL+C
 import os
 import sys
 import time
+import __builtin__
 from bcolors import bcolors as b
 
 class killswitch:
@@ -17,7 +18,7 @@ class killswitch:
 	def __init__(self):
 		pass
 
-	def killswitch(self, pid):
+	def skillswitch(self):
 		#
 		# determined this file needs to handle its own exceptions, no offloading
 		#
@@ -25,12 +26,20 @@ class killswitch:
 			try: 
 				time.sleep(0) # dirty hack to monitor for keybard interrupt
 			except KeyboardInterrupt:
-				os.popen("iptables -t nat -F")
-			if str(diag).lower() in ["true","yes"]:
-				print "Killer module killed pid: " + str(pid)
+				if str(sr1) in ["client","c"]:
+					f.close()
+					print b.OKBLUE+"\t[!] User exited - closed log file ...\n" + b.ENDC
+					sys.exit(0)
+				if str(sr1) in ["server","s"]:
+					os.popen("iptables -t nat -F")
+					if str(diag).lower() in ["true","yes"]:
+						print "Killer module killed pid: " + str(pid)
+					print b.OKBLUE+"\t[!] User exited - flushed IPTABLES and killed server pid "+str(pid)+" ...\n" + b.ENDC
+					pid = os.getpid()
+					killer = "kill -9 "+str(pid)
+					os.popen(killer)
+					sys.exit(0)
 		except Exception as killerdied:
 			print "exception caused in killswitch module: "+str(killerdied)
-		print b.OKBLUE+"\t[!] User exited - flushed IPTABLES and killed server pid "+str(pid)+" ...\n" + b.ENDC
-		killer = "kill -9 "+str(pid)
-		os.popen(killer)
-		sys.exit(0)
+			sys.exit(0)
+
